+++ lib/beam_file/normalizer.ex
@@ -103,8 +103,13 @@ defmodule BeamFile.Normalizer do
     end
   end
 
-  def normalize({{:., meta1, [:erlang, :binary_to_atom]}, meta2, [arg, :utf8]}, _target) do
-    {{:., meta1, [:erlang, :binary_to_atom]}, meta2, [arg, {:__block__, [], [:utf8]}]}
+  if Version.match?(System.version(), "~> 1.13.0") do
+    # In Elixir 1.13 the original AST throws an error in the formatter.
+    # ** (CaseClauseError) no case clause matching: :utf8
+    # :utf8 is replaced with the equivalent AST {:__block__, [], [:utf8]}
+    def normalize({{:., meta1, [:erlang, :binary_to_atom]}, meta2, [arg, :utf8]}, :code) do
+      {{:., meta1, [:erlang, :binary_to_atom]}, meta2, [arg, {:__block__, [], [:utf8]}]}
+    end
   end
 
   def normalize({:for, meta, args}, target) when is_list(args) do